% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{misclass_err}
\alias{misclass_err}
\title{calculate classification error}
\usage{
misclass_err(pred, actual)
}
\arguments{
\item{pred}{predicted values by the model.}

\item{actual}{actual values.}
}
\value{
Return the classification error
}
\description{
calculate classification error
}
